test(daemon): isolate umask-mutating socket tests in their own process
`bind_private` (#995) tightens the process-global umask to `0o177` for
the synchronous span of its socket `bind`. That is safe in production
(a one-shot startup bind with no concurrent file creation), but umask
is shared across every thread in the process. The library's unit-test
binary runs thousands of tests in parallel, so a daemon test's bind
window races any other test creating a file in the same instant: the
new inode loses permission bits and that test fails with `EACCES`.

Issue #1017 saw it flake the `bind_private` owner-only-socket test, but
the race is binary-wide: under load it also struck `data::`, `git::`,
and `datadog::` temp-file tests, which is what reddened CI.

Keep the umask hardening untouched and remove the race instead: move
the umask-mutating tests (the `bind_private` check and the `server`
tests that bind via `bind_or_reclaim`/`run`) into a dedicated
integration-test binary, `tests/daemon_socket.rs`. A separate test
binary is a separate process, and umask is per-process, so its bind
windows can no longer touch the unit tests. Within that binary the few
tests restore their own temp dir to `0700` before binding, guarding the
intra-binary race. `bind_private` is exposed `#[doc(hidden)] pub` so
the integration test can exercise it directly.

Verified: 14 full `cargo test --all-features` runs with zero umask
`EACCES` failures (previously a failure on every CI job); clippy
`--all-targets` clean; fmt clean.

Closes #1017
