+++ src/daemon/server.rs
@@ -176,108 +176,7 @@ pub fn resolve_socket(socket: Option<PathBuf>) -> Result<PathBuf> {
     }
 }
 
-#[cfg(test)]
-#[allow(clippy::unwrap_used, clippy::expect_used)]
-mod tests {
-    use super::*;
-    use crate::daemon::services::echo::EchoService;
-    use serde_json::Value;
-    use std::time::Duration;
-
-    /// Spins up a daemon on a temp socket, runs `body`, then shuts it down and
-    /// joins the server task.
-    async fn with_daemon<F, Fut>(body: F)
-    where
-        F: FnOnce(PathBuf) -> Fut,
-        Fut: std::future::Future<Output = ()>,
-    {
-        let dir = tempfile::tempdir().unwrap();
-        let socket = dir.path().join("d.sock");
-        let mut registry = ServiceRegistry::new();
-        registry.register(Arc::new(EchoService));
-        let opts = DaemonOptions {
-            socket_path: socket.clone(),
-        };
-        let handle = tokio::spawn(run(registry, opts));
-
-        // Wait for the socket to accept.
-        let client = crate::daemon::client::DaemonClient::new(&socket);
-        for _ in 0..50 {
-            if client.ping().await.is_ok() {
-                break;
-            }
-            tokio::time::sleep(Duration::from_millis(20)).await;
-        }
-
-        body(socket.clone()).await;
-
-        client.shutdown().await.ok();
-        let _ = tokio::time::timeout(Duration::from_secs(2), handle).await;
-    }
-
-    #[tokio::test]
-    async fn ping_and_status_and_routing() {
-        with_daemon(|socket| async move {
-            let client = crate::daemon::client::DaemonClient::new(&socket);
-
-            // Built-in ping.
-            client.ping().await.unwrap();
-
-            // Built-in status reports the echo service.
-            let report = client.status().await.unwrap();
-            assert_eq!(report.services.len(), 1);
-            assert_eq!(report.services[0].name, "echo");
-            assert!(report.services[0].healthy);
-
-            // Routed op reaches the echo service and round-trips the payload.
-            let reply = client
-                .request(DaemonEnvelope::service(
-                    "echo",
-                    "echo",
-                    json!({ "hello": "world" }),
-                ))
-                .await
-                .unwrap();
-            assert!(reply.ok);
-            assert_eq!(reply.payload, json!({ "hello": "world" }));
-
-            // Unknown service is an error, not a panic.
-            let reply = client
-                .request(DaemonEnvelope::service("nope", "x", Value::Null))
-                .await
-                .unwrap();
-            assert!(!reply.ok);
-            assert!(reply.error.unwrap().contains("unknown service"));
-
-            // Unknown built-in op is an error.
-            let reply = client
-                .request(DaemonEnvelope::builtin("frobnicate"))
-                .await
-                .unwrap();
-            assert!(!reply.ok);
-        })
-        .await;
-    }
-
-    #[tokio::test]
-    async fn second_bind_is_refused_while_first_is_live() {
-        with_daemon(|socket| async move {
-            let mut registry = ServiceRegistry::new();
-            registry.register(Arc::new(EchoService));
-            let err = single_instance::bind_or_reclaim(&socket).await.unwrap_err();
-            assert!(err.to_string().contains("already running"));
-        })
-        .await;
-    }
-
-    #[tokio::test]
-    async fn stale_socket_is_reclaimed() {
-        let dir = tempfile::tempdir().unwrap();
-        let socket = dir.path().join("d.sock");
-        // A leftover regular file at the socket path stands in for a stale
-        // socket: nothing is listening, so the ping probe fails and we reclaim.
-        std::fs::write(&socket, b"stale").unwrap();
-        let listener = single_instance::bind_or_reclaim(&socket).await.unwrap();
-        drop(listener);
-    }
-}
+// The daemon-server tests that bind a socket (and thus mutate the process-global
+// umask via `bind_or_reclaim`) live in `tests/daemon_socket.rs`, isolated in
+// their own process so the umask write cannot race the library's other parallel
+// unit tests. See #1017.
