+++ src/daemon/single_instance.rs
@@ -63,7 +63,14 @@ pub async fn bind_or_reclaim(path: &Path) -> Result<UnixListener> {
 /// so no other task on this worker thread can observe the tightened umask;
 /// genuinely-parallel OS threads creating files in the same instant are the only
 /// residual exposure, which is acceptable for a one-shot startup bind.
-fn bind_private(path: &Path) -> std::io::Result<UnixListener> {
+///
+/// Exposed (`#[doc(hidden)]`, not a stable API) so it can be exercised directly
+/// from `tests/daemon_socket.rs`. That process-global umask write would race
+/// any *other* parallel unit test creating a file in the same instant, so the
+/// tests that drive it live in their own integration-test binary — a separate
+/// process whose umask never touches the library's unit-test threads. See #1017.
+#[doc(hidden)]
+pub fn bind_private(path: &Path) -> std::io::Result<UnixListener> {
     #[cfg(unix)]
     let _umask = UmaskGuard::set(nix::sys::stat::Mode::from_bits_truncate(0o177));
     UnixListener::bind(path)
@@ -88,22 +95,7 @@ impl Drop for UmaskGuard {
     }
 }
 
-#[cfg(all(test, unix))]
-#[allow(clippy::unwrap_used, clippy::expect_used)]
-mod tests {
-    use super::*;
-    use std::os::unix::fs::PermissionsExt;
-
-    /// `bind_private` alone — with no follow-up `chmod` — must yield a `0600`
-    /// socket, proving the umask closes the window rather than a post-bind
-    /// `set_file_0600`. Needs a Tokio runtime to register the listener fd.
-    #[tokio::test]
-    async fn bind_private_creates_an_owner_only_socket() {
-        let dir = tempfile::tempdir().unwrap();
-        let socket = dir.path().join("d.sock");
-        let listener = bind_private(&socket).unwrap();
-        let mode = std::fs::metadata(&socket).unwrap().permissions().mode() & 0o777;
-        assert_eq!(mode, 0o600, "socket mode was {mode:o}, expected 600");
-        drop(listener);
-    }
-}
+// The umask-driven `bind_private`/`bind_or_reclaim` tests live in
+// `tests/daemon_socket.rs`: their process-global umask write would race the
+// library's other parallel unit tests, so they run in a separate process. See
+// #1017.
